requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
discord.py==2.3.2
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
setuptools==68.1.2
soupsieve==2.5
urllib3==2.0.5
wheel==0.41.2
yarl==1.9.2
environs==9.5.0
SQLAlchemy==2.0.23
uvicorn==0.25.0
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic~=2.5.3
fastapi~=0.109.0
starlette-admin==0.13.2
babel==2.14.0
python-jose==3.3.0
adaptix==3.0.0b5
itsdangerous==2.1.2
aio_pika==9.4.1
faststream==0.5.2